Clean up serialisation of eas_email_info.
There's no need to precalculate the buffer size. If it was an optimisation,
that would be one thing. But if we need to carefully precalculate it and
keep the size calculation in sync with what *actually* goes into it, that
doesn't seem worth it. Just use a GString to build it up.