Reworked how eas_connection_send_request() deals with the 'doc' object over its lifetime in for both AS 14.0 and 12.1.
The special case for SendMail should also be handled, and the
send mail request and message GObjects are updated to pass ownership
to eas_connection_send_request() like everything else.

All of this will need to be refactored based around the lifetime
of the message to the server only expiring when we know the message
doesn't need to be resent.