syncevolution: hard-code libeassync in configure instead of Makefile
The Makefile.am should be "clean" without hacks because it is meant to
be used also when compiling within SyncEvolution. Therefore this
commit moves the hard-coding of libeassync flags into the activesyncd
configure.

It also renames the variables such that they are consistent with the
result of PKG_CONFIG_CHECK(): upper case, no lib prefix, _LIBS instead
of _LIBADD.