Bump WindowSize back up to 100 now that the timeout problem is fixed
Not sure it's ideal, but I think it's slightly better than 25. Testing
with the Intel servers, the server seems to take about 6 seconds of
pondering before even *starting* to respond to a sync request with
WindowSize of 25. But it only ponders for 18 seconds for WindowSize 100.
So assuming the actual download time is unchanged for the bulk transfer,
that's a 25% reduction in time-per-message.