Add a test client library that allows a test provide a list of files that hold XML body responses to be set into the EasConnection GObject.
If there are entries in the list, EasConnection will swap the next
response from the server for the one in the file provided, after
which the file is removed from the list.

Also enabled a environment variable 'EAS_CAPTURE_RESPONSE'
If this is set then all decoded messages from the server will be
written to ~/eas-debug/<sequence number>.<process id>.xml

Messages from the server which cause an error will always be written
to file, in their WBXML format if that is where the problem arises.