Revert "changing handling of default folders to reside in daemon rather than client"
This reverts commit 35963bb8e086784dc21589ab13abcc0ede58e512.

It breaks Sync message handling, causing the dæmon to send *two* Sync
messages in sequence and discard the results from the first.

On loading up Evolution for the first time with a large mailbox, this
results in only half your messages (every other hundred) being shown.
And on using it in normal 'incremental' operation it results in *no* new
messages being shown.

Not bothering to diagnose and fix; the commit is too large (I'd probably
have made it at least three separate commits to make it easy to follow),
and it was *easily* tested but was committed broken anyway.

So I'm just grumpily reverting it.