Revert "fixing add calendar request to use default folders, and patching timezone stuff out because it doesn't work - chore C81 for sorting this"
This reverts commit 18004a4fde616ec987e585b5c5050e9390ac1ba8.

There was nothing obviously wrong with this, but I already had to revert
commit 35963bb8e086784dc21589ab13abcc0ede58e512 because *that* was broken.

This commit depends on parts of that other reverted commit. It
*shouldn't* have done; commit 35963bb8 should have been about three
separate commits, and this one would have depended on *one* of those,
while I'd have needed to revert a *different* one.

But since 35963bb8 wasn't committed in bite-sized chunks as it should
have been, the only option is to revert the whole thing and also this
commit which depends on it.