Use eas_connection_get_account() in eas_sync_req_Activate()
There's no need to find our own EasAccount. I'll remove priv->acc shortly.

Note that I can't use priv->cnc here either, because it's broken. It *looks*
like it's set in eas_sync_req_new(), but it isn't, because at that point
our caller won't have set it in the EasRequestBase yet. So priv->cnc is
always NULL. And there's no point in fixing it because there are no other
users anyway; we might as well get the connection only once, if we need it.