Update own version of server_flags when pushing changes to server
Now that we use the sync_key returned by the server in update_emails, we no
longer use an *old* sync key in our next communication to the server. This
means that the server no longers assumes that we crashed, and no longer
*tells* us about the changes that we made. So we need to update the
server_flags field in the message_info, to prevent us from repeatedly
sending the *same* changes to the server.