Save summary db after clearing invalid sync key
By hitting Ctrl-C, I managed to quit Evolution leaving it in a state
where I had cleared the folder, but *hadn't* written the "0" sync_key out
to the summary db. I noticed it had cleared the folder, and killed it before
the first fetch request had finished.

Make it save the folder summary db immediately... in fact, notice the the
tail of the while() loop is doing just that, and some more stuff that we
should probably be doing, and refactor the code to take that route instead
of doing a 'continue'.