Remove priv->connection from EasMail object. It's wrong.
We don't have an EasMail object per account; we have *one* object. So
storing the connection pointer in its private data is inappropriate.

What's more, it's completely unnecessary. In all cases it can be a
local variable in the function that wants to know it, which was always
looking it up for itself (correctly) anyway.