Revert "remove common object as it is not used"
This reverts commit 244b299bce135d31f9474ce78e7299a829834ff7. You're right;
it's not used. But I was about to start using it... :)

We'll need it for operations on the folder tree (which the dæmon has to
handle centrally rather than just exposing sync_folder_hierarchy to each
client and letting them stomp over each others' sync keys.

I was also planning to put a generic bidirectional Sync call into it,
for use with *all* object types. And GetItemEstimate too.