ActiveSyncSource.cpp: allow retrieval of items beyond the single-call window.
eas_sync_handler_get_items has a window (currently set at 100 items),
and a "more_available" value that's returned if there are more items
available than it's able to retrieve in one call. This patch uses this
value and loops the _get_items function as long as there are items.