Ensure server responses are valid UTF-8 by substituting � for bad bytes.
Yet another iteration of our attempts to cope with the fact that Exchange returns
XML with invalid UTF-8 bytes in it. There were two problems with the current code.

Firstly, the trick of setting XML_PARSE_RECOVER wasn't sufficient; it was causing
weirdness with expansion of character elements like &quot; later in the XML.

Secondly, if the invalid UTF-8 gets into an actual DBus response, the dbus dæmon
will instantly disconnect us and then dbus-glib will exit the program immediately
without any hint about why (which was fun to debug).

That second observation leads me to believe that I won't be breaking anything
by doing this; the only data which *need* to be legacy 8-bit crap are the
contents of certain emails, and those are handled as separate parts of a
multipart fetch (and stored in a file rather than being passed back over DBus,
thankfully).

Hopefully this is the last we hear of it...