Increase priority of deferred SoupMessage cancellation.
There is weirdness here, and I suspect we need to be bumping the refcount on
 the objects which will are passed to call_soup_session_cancel_message().
What happens if the cancellation gets scheduled just as the last block of
data comes in, and call_soup_session_cancel_message() happens *after* the
request has actually completed successfully?

This is not a real fix, but it's probably a step in the right direction.
It stops us from continuing to happily receive data on a message that
should already have been cancelled.

It *sucks* that libsoup isn't thread-safe and has to be called from its
own thread.