Handle arrays in DBus result.
g_variant_is_of_type() won't match a result of (sas) when the format string
is (s^as), which seems rather stupid to me. So it *can't* be used as a general
"is it valid to call g_variant_get() on this GVariant?" function.

We have to work around it by stripping the ^ characters from the string
we use for checking.

Not entirely pretty, but it's either this or take two almost-identical
type/format strings as arguments to eas_gdbus_mail_call().