calendar: support detached recurrences without parent (BMC #22831)
The ActiveSync data model requires that exceptions have a parent event
with a recurrence rule that matches all exceptions. This patch
synthesizes such an event so that it is never visible to the user
because all recurrences are overridden by the detached recurrences or
additional EXDATEs.

The difficulty in this approach (and its complexity) lies in guessing
the right recurrence rule. For a single event, any recurrence rule
would work. Yearly recurrence is currently tried first and used. For
multiple events, rules are synthesized starting with yearly
recurrence, continuing with monthly, etc. The interval between the
first two detached recurrences is also tried. An rrule works if it
includes all known detached recurrences. If it recurrs more often,
additional EXDATEs are needed (there the rules with less repitions
are tried first).

If the wrong recurrence rule was chosen, with wrong in the sense of
"not what the real parent event had", then it might be impossible to
import additional detached recurrences because they won't match the
currently stored rule of the artificial parent. Need to test... which
might not be easy. I couldn't convince Outlook Web Access to import
multiple detached recurrences at all.

The alternative would be to pretend that the detached recurrences are
independent events. That leads to problems with UID (should better be
changed) and RECURRENCE-ID (cannot be sent). Therefore importing an
update for a single detached recurrence would not work.