calendar: updated time zone sending
Exchange 2010 on 123together.com added its own time zone when
an event was sent without an explicit time zone. This is incorrect
for events which were meant to be in UTC time zone, so now add
the UTC time zone definition to those.

Also send a "better" definition of the UTC time zone, meaning that
it is not completely empty as before. However, that still didn't
help in some cases: meetings without recurrence still end up with
that UTC time zone getting replaced.