calendar: added all day event time conversion heuristic
The meaning of "all day" is poorly specified. Besides setting
the flag it also seems to be necessary to align the UTC start/end
times so that they align with midnight in the desired taget time
zone. This is how Exchange sends all day events created via OWA.

However, Exchange 2010 on 123together.com replaces the time zone
definition of all day events created on the server via ActiveSync
*without* updating the time, so that the alignment with midnight
becomes broken doing the time conversion can shift the days of the
events.

Workaround is to double-check that the converted times align with
midnight, otherwise skipping the time conversion.