logging: set handler only in activesyncd executable
The libraries should not set output handlers and instead leave
message routing to the executable in which they are used. SyncEvolution
uses that to put the library messages into its own log files.

The activesyncd continues to use the same log handler as before,
so that EAS_DEBUG_FILE continues to work.