remove usage of libdbus
libdbus was still checked for in configure, added to compiler and linker
invocations, included in source code and even called directly at the
shutdown of the daemon (dbus_shutdown()).

This commit removes all of these remaining traces of libdbus. The main
motivation is that dbus_shutdown() called exit() with a non-zero
result code, and therefore it was impossible to determine whether the
server had shut down normally.

Another motivation is that this change will make it harder to "accidentally"
use libdbus instead of GIO GDBus, because the linker on modern Linux
distros will warn about calling functions indirectly (= without linking
directly to the lib that provides them).