Start porting to 3.4.
This is incomplete; I'm not going to get this building before I go away for
the week. Commit what I have, in the vain hope that I can convince someone
else to keep going.

The rest of what needs doing will be extremely similar to commit f7dd1cd2
and its parent 8cf26212 in evolution-ews.