activesyncd: fix signal handler
Sometimes in SyncEvolution's nightly testing the activesyncd dies
with a crash after being sent a SIGTERM and a SIGINT (both are
sent because the sending script is used for more than activesyncd
and doesn't know which one will work).

The debug output in that case shows:

(process:20189): activesyncd-DEBUG (recursed): signalHandler++

(process:20189): activesyncd-DEBUG (recursed): signalHandler++

...

Apparently at some point the process runs out of memory, or gets
killed with SIGKILL after not shutting down in a timely manner (don't
remember).

This patch solves several issues with signal handler and seems to help:
- Don't produce IO in the signal because it is uncertain
  whether that is allowed.
- Don't leave a dangling GMainLoop pointer in g_mainloop after
  freeing it (probably the reason for the crash).
- Avoid a race condition by clearing g_mainloop before freeing it.

I also think catching SIGABRT is a bad idea because if something
catastrophic happens and raises that signal, the process should
shut down abnormally. I left it in place for now.