libeasclient: add eas_sync_handler_get_folder_list
Folder listing code was moved into eas-folder.c and the existing
eas_mail_handler_get_folder_list was modified to use the new common code.
A new function eas_sync_handler_get_folder_list (with the same parameters and
behaviour, except for requiring a sync handle instead of a mail handle) was
added for the convenience of sync clients. This will be used by libeasclient
and the syncevolution ActiveSync backend.