eas-daemon: workround folder sync when server loses state
This is a hack to workround the problem that the ActiveSync server can lose
state and demand that we restart the folder sync with key 0.  The hack is to
**always** do every folder sync starting from key 0.  Of course, this means
that if there are large numbers of folders or changes are rare then extra data
has to be communicated.

A more subtle fix would be to only reset the sync if the server complains about
an invalid sync key.  That could be considered for a future update.