restore compatibility with EDS 3.4
EDS 3.4 does not yet have the unified header files. We can (and should) only
include them if the version check from eds-version.h (which is always
available and usable directly) tells us that EDS is recent enough.

A compile-time check is used instead of a configure check to avoid making
the header files depend on config.h and a specific check, because that might
not work if the header files are used inside a client app. Not sure whether
that is ever the case - better safe than sorry.