calendar: fix VEVENT->EAS conversion of multi-day events
All-day events stretching over multiple days were not marked as AllDayEvent
when sending via ActiveSync. This did not matter with Exchange 2010, which
automagically set the flag for us and thus let the tests in SyncEvolution
pass.

Exchange 2010 SP2 no longer does that, IMHO correctly, thus the tests started
to fail after switching to a test account on Exchange 2010 SP2.

The reason for the missing AllDayEvent flag was an explicit test that start
and end time are SECONDS_PER_DAY apart. That was probably done to comply with
the "end of next day" part of the AllDayEvent spec, but is not needed
in practice and also wrong (because not all days have the same number
of seconds - think daylight saving time changes). Removing the check fixes
the problem (only tested with Exchange 2010 SP2, though).