autotools: avoid hard dependency on Evolution dev files
PKG_CHECK_MODULES() aborted when it didn't find some .pc files, even
if those files where not needed at all because the feature depending
on them was disabled.

Now the files are checked and the result is remembered instead of
aborting. Then the feature check for eplugin is enabled by default
only if found, and fails hard if enabled explicitly without the .pc.

The feature check for camel should be updated separately, taking into
account/fixing that it no longer compiles with libcamel from EDS 3.12.

Compilation with EDS 3.12 works with --disable-camel-backend
--disable-qtconfig --disable-eplugin.