_eas2ical_convert_datetime_property: if time for all day event fails sanity check (not midnight), do not use "date only" time format.
In some circumstances Exchange supplies the wrong timezone information for an
all day event, so the apparent local time of the event is not midnight. In that
case there is little we can do -- we do not know which way the time should be
adjusted, so we do not know on which day the event was intended to fall.

So, in that case, we include the time in the data (not just the date) in the
hopes that this provides the user with more information to determine the
correct day.  In most software, this results in the event not being treated
as an "all day" event but that is better than appearing as an all day event on
the wrong day.