autotools: fix --disable-camel-backend
--disable-camel-backend is necessary to compile activesyncd on
distros without the recent backend API. Such a reduced activesyncd
is still useful for contact and calendar syncing.

To fix the option, entering and thus compiling source directories which are
only needed for the backend must be avoided when --disable-camel-backend is
used. This broke when adapting the code to API changes in EDS.